checks if there is an attribute with prefix short or long (e.g. shortTitle or longAbstract) as well as "long" is now the default form
in specs: variable - renders the text contents of a variable. Attribute value must be one of the standard variables. May be accompanied by the form attribute to select the “long” (default) or “short” form of a variable (e.g. the full or short title). If the “short” form is selected but unavailable, the “long” form is rendered instead.