Fix issue when rendering date ranges
When rendering a date range, its possible that the range doesn't
need to be rendered at all. For example is we are rendering the
year and we have a range for the day. In that case we can just
render the date as if it wasn't a range at all.